requirements.txt
aniso8601==9.0.1
click==8.1.7
Flask==3.1.0
Flask-HTTPAuth==4.5.0
Flask-RESTful==0.3.9
itsdangerous>=2.2
MarkupSafe>=2.1.1
pytz==2021.3
six==1.16.0
pandas==2.2.3
marshmallow==4.0.0
flask-marshmallow==1.3.0
flask-cors==6.0.0
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1